pybricks: MicroPython on EV3RT: proof of concept
This app allows you to run a Python user script that is stored on the microSD card.

This is based on the helloev3 example with MicroPython linked as a library. The MicroPython HAL additions are somewhat hacky and still very minimal, but this might serve as a useful starting point.