pybricks: periodic poll task and mutex demo
This implements a poll task that runs at fixed intervals. It does not run if a user currently claims the shared resource. On the contrary the user call always waits for the poll task to complete before running.

The reasoning is as follows -- asuming we have only one motor.

The poll task runs to update control tasks previously activated by a pbio_motor_run_ call. It does not run if a new command is currently being processed. This is fine because the new command call synchronously calls the control update.

This is different from what we currently have in Pybricks; so for experimental purposes only; we could also just implement it like we have it on the other hubs.